Eureka: move DataCenterInfo to EurekaConnection
In the Java implementation, InstanceInfo is best thought of as a utility
class, used only for marshalling/unmarshalling to JSON. The Java also
uses it as a mutable place to keep "status" (are we up or down right
now). Because it can be purely computed based on the other things in a
EurekaConnection + DataCenterInfo, there's no reason to make it part of
the record. Delete it for now; we'll reintroduce it shortly.

While we're here, it seems obvious that we should export some
DataCenterInfo stuff, so do that too.