Eureka: implement lookupByAppName
The response format for looking up a single application is kind of
wacky: it's an "application" response, wrapped in a toplevel
{"application": ...}. An "application" response has an "instance" field,
which can be an object (if there's only one instance up) or a list (if
there are several). Cope with this.

Make some other changes to the InstanceInfo and DataCenterInfo parsers
to cope with optional fields and wacky formatting. Specifically, "port"
and "securePort" are serialized as

    "port": {"@enabled": false, "$": "80"}