Types: home page URL is optional
It turns out that it's possible to write a client that causes Eureka to
not have a homePageUrl for a service, which means that we have to
represent this using a Maybe.

It turns out that the person who wrote the client that causes Eureka to
not have a homePageUrl is me and you're looking at the commit log of the
client.